Cleanup old auraescript and rename deno based code
