+++ auraescript/Cargo.toml
@@ -1,69 +1,25 @@
-# ---------------------------------------------------------------------------- #
-#             Apache 2.0 License Copyright © 2022 The Aurae Authors            #
-#                                                                              #
-#                +--------------------------------------------+                #
-#                |   █████╗ ██╗   ██╗██████╗  █████╗ ███████╗ |                #
-#                |  ██╔══██╗██║   ██║██╔══██╗██╔══██╗██╔════╝ |                #
-#                |  ███████║██║   ██║██████╔╝███████║█████╗   |                #
-#                |  ██╔══██║██║   ██║██╔══██╗██╔══██║██╔══╝   |                #
-#                |  ██║  ██║╚██████╔╝██║  ██║██║  ██║███████╗ |                #
-#                |  ╚═╝  ╚═╝ ╚═════╝ ╚═╝  ╚═╝╚═╝  ╚═╝╚══════╝ |                #
-#                +--------------------------------------------+                #
-#                                                                              #
-#                         Distributed Systems Runtime                          #
-#                                                                              #
-# ---------------------------------------------------------------------------- #
-#                                                                              #
-#   Licensed under the Apache License, Version 2.0 (the "License");            #
-#   you may not use this file except in compliance with the License.           #
-#   You may obtain a copy of the License at                                    #
-#                                                                              #
-#       http://www.apache.org/licenses/LICENSE-2.0                             #
-#                                                                              #
-#   Unless required by applicable law or agreed to in writing, software        #
-#   distributed under the License is distributed on an "AS IS" BASIS,          #
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.   #
-#   See the License for the specific language governing permissions and        #
-#   limitations under the License.                                             #                                                                             #
-#                                                                              #
-# ---------------------------------------------------------------------------- #
-
 [package]
 name = "auraescript"
 version = "0.1.0"
 edition = "2021"
-authors = ["The Aurae Authors", "Kris Nóva <kris@nivenly.com>"]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [[bin]]
 name = "auraescript"
 path = "src/bin/main.rs"
 
 [dependencies]
-rhai = { version = "1.9.1", features = ["metadata"] }
-tonic = { version = "0.8", features = ["tls"] }
-tokio-stream = { version = "0.1", features = ["net"] }
-tokio = { version = "1.0", features = ["macros", "fs", "rt-multi-thread"] }
-toml = "0.5.9"
-serde = { version = "1.0.144", features = ["derive"] }
+anyhow = "1.0.66"
+deno_ast = { version = "0.21.0", features = ["transpiling"] }
+deno_core = "0.160.0"
+macros = { package = "auraescript_macros", path = "../auraescript_macros" }
+prost = "0.11.2"
+serde = "1.0.147"
+tokio = { version = "1.22.0", features = ["fs", "rt-multi-thread"] }
+tonic = { version = "0.8.2", features = ["tls"] }
 tower = "0.4.13"
-home = "0.5.3"
-anyhow = "1.0"
-x509-parser = "0.14.0"
-rustls-pemfile = "1.0.1"
-if_chain = "1.0.2"
-time = "0.3.14"
-oid-registry = "0.6.0"
-x509-certificate = "0.14.0"
-prost = "0.11"
-ring = "0.16.20"
-json = "0.12.4"
-serde_json = "1.0.85"
-macros = { path = "macros" }
+x509-certificate = "0.15.0"
+toml = "0.5.9"
 
 [build-dependencies]
-tonic-build = "0.8"
-anyhow = "1.0"
-
-[dev-dependencies]
-openssl = "0.10.41"
+anyhow = "1.0.66"
+tonic-build = "0.8.2"
