+++ auraescript/build.rs
@@ -28,12 +28,58 @@
  *                                                                            *
 \* -------------------------------------------------------------------------- */
 
-use anyhow::Result;
-//use std::process::Command;
+use anyhow::{anyhow, Result};
+use std::path::PathBuf;
+use std::process::Command;
 
 fn main() -> Result<()> {
-    // Example running "make command" during the build
-    //Command::new("make").args(&["command"]).status().unwrap();
+    let out_dir = match std::env::var("CARGO_MANIFEST_DIR") {
+        Ok(out_dir) => {
+            let mut out_dir = PathBuf::from(out_dir);
+            out_dir.push("lib");
+            out_dir
+        }
+        _ => PathBuf::from("lib"),
+    };
+
+    println!("cargo:rerun-if-changed=\"src\"");
+    println!("cargo:rerun-if-changed=\"{:?}\"", out_dir);
+
+    // https://github.com/stephenh/ts-proto
+    match Command::new("npm").args(["install", "ts-proto"]).status()?.code() {
+        Some(0) => {}
+        _ => {
+            return Err(anyhow!(
+                "Command `npm install ts-proto` failed. Is npm installed?"
+            ));
+        }
+    }
+
+    let proto_message_files = ["../api/v0/runtime.proto"];
+
+    for file in proto_message_files {
+        println!("cargo:rerun-if-changed=\"{file}\"");
+
+        let status = Command::new("protoc")
+            .args([
+                "--plugin=./node_modules/.bin/protoc-gen-ts_proto",
+                &format!("--ts_proto_out={}", out_dir.display()),
+                "--ts_proto_opt=outputEncodeMethods=false",
+                "--ts_proto_opt=outputClientImpl=false",
+                "-I=../api/v0",
+                file,
+            ])
+            .status()?;
+
+        match status.code() {
+            Some(0) => {}
+            _ => {
+                return Err(anyhow!(
+                    "Failed to generate Typescript file for proto '{file}'"
+                ))
+            }
+        }
+    }
 
     generate_grpc_code()?;
 
@@ -44,10 +90,14 @@ fn generate_grpc_code() -> Result<()> {
     let mut tonic_builder = tonic_build::configure();
 
     // Generated services use unwrap. Add them here to suppress the warning.
-    for service in ["meta", "observe", "runtime", "schedule"] {
+    for service in ["observe", "runtime", "schedule"] {
         tonic_builder =
             tonic_builder.server_attribute(service, "#[allow(missing_docs)]");
     }
+
+    // Add proto messages here to add `#[derive(::serde::Serialize, ::serde::Deserialize)]` to the type.
+    // If other code generation needs these derive attributes, they will automatically be added.
+    let mut derive_serde_serialize_deserialize: Vec<&str> = vec![];
 
     // Types generated from proto messages derive PartialEq without Eq. Add them here to suppress the warning.
     for message in [
@@ -62,122 +112,28 @@ fn generate_grpc_code() -> Result<()> {
         "runtime.StartCellResponse",
         "runtime.StopCellRequest",
         "runtime.StopCellResponse",
-        "observe.GetAuraeDaemonLogStreamRequest",
-        "observe.GetSubProcessStreamRequest",
-        "observe.LogItem",
     ] {
         tonic_builder = tonic_builder.type_attribute(
             message,
             "#[allow(clippy::derive_partial_eq_without_eq)]",
         );
+
+        derive_serde_serialize_deserialize.push(message);
     }
-
-    // Add proto messages here to add `#[derive(::serde::Serialize, ::serde::Deserialize)]` to the type.
-    // If other code generation (e.g., `::macros::Output`) needs these derive attributes, they will automatically be added.
-    let mut derive_serde_serialize_deserialize: Vec<&str> = vec![];
-
-    // Add proto messages here to generate output functions (e.g., `json()`) on the type
-    for message in ["runtime.Cell", "runtime.Executable", "observe.LogItem"] {
-        if !derive_serde_serialize_deserialize.contains(&message) {
-            derive_serde_serialize_deserialize.push(message);
-        }
-
-        tonic_builder = tonic_builder
-            .type_attribute(message, "#[derive(::macros::Output)]");
-    }
-
-    // Add proto messages  here to generate getters and setters.
-    tonic_builder = generate_grpc_code_for_getters_setters(
-        tonic_builder,
-        vec![GetSet {
-            message: "runtime.Executable",
-            ignore_fields: vec!["meta"],
-            ..Default::default()
-        }],
-    );
 
     // Loop to add serde attributes
     for message in derive_serde_serialize_deserialize {
-        tonic_builder = tonic_builder.type_attribute(
-            message,
-            "#[derive(::serde::Serialize, ::serde::Deserialize)]",
-        )
-    }
-
-    tonic_builder.build_server(false).compile(
-        &[
-            "../api/v0/runtime.proto",
-            "../api/v0/observe.proto",
-            "../api/v0/schedule.proto",
-        ],
-        &["../api/v0"],
-    )?;
-
-    Ok(())
-}
-
-#[allow(clippy::single_element_loop)]
-fn generate_grpc_code_for_getters_setters(
-    mut tonic_builder: tonic_build::Builder,
-    message_settings: Vec<GetSet>,
-) -> tonic_build::Builder {
-    // Add proto messages here to generate getters and setters.
-    for getset in message_settings {
-        let attribute = match (getset.getters, getset.setters) {
-            (true, true) => "#[derive(::macros::Getters, ::macros::Setters)]",
-            (true, false) => "#[derive(::macros::Getters)]",
-            (false, true) => "#[derive(::macros::Setters)]",
-            (false, false) => continue,
-        };
-
-        tonic_builder = tonic_builder.type_attribute(getset.message, attribute);
-
-        for field in getset.ignore_fields {
-            let path = format!("{}.{}", getset.message, field);
-            tonic_builder =
-                tonic_builder.field_attribute(path, "#[getset(ignore)]");
-        }
-
-        for field in getset.getters_ignore_fields {
-            let path = format!("{}.{}", getset.message, field);
-            tonic_builder =
-                tonic_builder.field_attribute(path, "#[getset(ignore_get)]");
-        }
-
-        for field in getset.setters_ignore_fields {
-            let path = format!("{}.{}", getset.message, field);
-            tonic_builder =
-                tonic_builder.field_attribute(path, "#[getset(ignore_set)]");
-        }
+        tonic_builder = tonic_builder
+            .type_attribute(
+                message,
+                "#[derive(::serde::Serialize, ::serde::Deserialize)]",
+            )
+            .type_attribute(message, "#[serde(default)]");
     }
 
     tonic_builder
-}
+        .build_server(false)
+        .compile(&["../api/v0/runtime.proto"], &["../api/v0"])?;
 
-struct GetSet {
-    /// The name of the proto message
-    message: &'static str,
-    /// Field names to ignore (no get or set function)
-    ignore_fields: Vec<&'static str>,
-    /// Should getters be generated
-    getters: bool,
-    /// Field names to ignore only when generating getters
-    getters_ignore_fields: Vec<&'static str>,
-    /// Should setters be generated
-    setters: bool,
-    /// Field names to ignore only when generating setters
-    setters_ignore_fields: Vec<&'static str>,
-}
-
-impl Default for GetSet {
-    fn default() -> Self {
-        Self {
-            message: "",
-            ignore_fields: vec![],
-            getters: true,
-            getters_ignore_fields: vec![],
-            setters: true,
-            setters_ignore_fields: vec![],
-        }
-    }
+    Ok(())
 }
