+++ auraescript/examples/script.ts
