+++ auraescript/src/builtin/config.rs
@@ -30,8 +30,8 @@
 
 //! Configuration used to authenticate with a remote Aurae daemon.
 //!
-//! Follows an ordered priority for searching for configuration on a clients
-//! machine.
+//! [`AuraeConfig::try_default()`] follows an ordered priority for searching for
+//! configuration on a client's machine.
 //!
 //! 1. ${HOME}/.aurae/config
 //! 2. /etc/aurae/config
@@ -41,22 +41,14 @@ use anyhow::{anyhow, Context, Result};
 use serde::Deserialize;
 use std::fs::File;
 use std::io::prelude::*;
+use std::path::Path;
 use toml;
-
-/// Configuration for AuraeScript client
-#[derive(Debug, Clone, Deserialize)]
-pub struct AuraeConfig {
-    /// Authentication material
-    pub auth: Auth,
-    /// System configuration
-    pub system: System,
-}
 
 /// The system configuration for AuraeScript.
 ///
 /// Used to define settings for AuraeScript at runtime.
 #[derive(Debug, Clone, Deserialize)]
-pub struct System {
+pub(crate) struct SystemConfig {
     /// Socket to connect the client to.
     pub socket: String,
 }
@@ -67,7 +59,7 @@ pub struct System {
 /// Changing this material during a process will impact the currently
 /// running process.
 #[derive(Debug, Clone, Deserialize)]
-pub struct Auth {
+pub(crate) struct AuthConfig {
     /// The same CA certificate the server has.
     pub ca_crt: String,
     /// The unique client certificate signed by the server.
@@ -76,38 +68,49 @@ pub struct Auth {
     pub client_key: String,
 }
 
-/// Attempt to easy-load Aurae configuration from well-known locations.
-pub fn default_config() -> Result<AuraeConfig> {
-    let home =
-        std::env::var("HOME").expect("missing $HOME environmental variable");
-    let path = format!("{}/.aurae/config", home);
-    let res = parse_aurae_config(path);
-    if res.is_ok() {
-        return res;
-    }
-
-    let res = parse_aurae_config("/etc/aurae/config".into());
-    if res.is_ok() {
-        return res;
-    }
-
-    let res = parse_aurae_config("/var/lib/aurae/config".into());
-    if res.is_ok() {
-        return res;
-    }
-
-    Err(anyhow!("unable to find config file"))
+/// Configuration for AuraeScript client
+#[derive(Debug, Clone, Deserialize)]
+pub(crate) struct AuraeConfig {
+    /// Authentication material
+    pub auth: AuthConfig,
+    /// System configuration
+    pub system: SystemConfig,
 }
 
-/// Attempt to parse a config file into memory.
-pub fn parse_aurae_config(path: String) -> Result<AuraeConfig> {
-    let mut config_toml = String::new();
-    let mut file = File::open(&path)?;
+impl AuraeConfig {
+    /// Attempt to easy-load Aurae configuration from well-known locations.
+    pub fn try_default() -> Result<Self> {
+        let home = std::env::var("HOME")
+            .expect("missing $HOME environmental variable");
 
-    if file.read_to_string(&mut config_toml)
-           .with_context(|| "could not read AuraeConfig toml")? == 0 {
-        return Err(anyhow!("empty config"));
+        let search_paths = [
+            &format!("{home}/.aurae/config"),
+            "/etc/aurae/config",
+            "/var/lib/aurae/config",
+        ];
+
+        for path in search_paths {
+            if let Ok(config) = Self::parse_from_file(path) {
+                return Ok(config);
+            }
+        }
+
+        Err(anyhow!("unable to find config file"))
     }
 
-    Ok(toml::from_str(&config_toml)?)
+    /// Attempt to parse a config file into memory.
+    pub fn parse_from_file<P: AsRef<Path>>(path: P) -> Result<AuraeConfig> {
+        let mut config_toml = String::new();
+        let mut file = File::open(path)?;
+
+        if file
+            .read_to_string(&mut config_toml)
+            .with_context(|| "could not read AuraeConfig toml")?
+            == 0
+        {
+            return Err(anyhow!("empty config"));
+        }
+
+        Ok(toml::from_str(&config_toml)?)
+    }
 }
