--- auraescript_deno/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "auraescript_deno"
-version = "0.1.0"
-edition = "2021"
-
-[[bin]]
-name = "auraescript"
-path = "src/bin/main.rs"
-
-[dependencies]
-anyhow = "1.0.66"
-deno_ast = { version = "0.21.0", features = ["transpiling"] }
-deno_core = "0.160.0"
-macros = { package = "auraescript_deno_macros", path = "../auraescript_deno_macros" }
-prost = "0.11.2"
-serde = "1.0.147"
-tokio = { version = "1.22.0", features = ["fs", "rt-multi-thread"] }
-tonic = { version = "0.8.2", features = ["tls"] }
-tower = "0.4.13"
-x509-certificate = "0.15.0"
-toml = "0.5.9"
-
-[build-dependencies]
-anyhow = "1.0.66"
-tonic-build = "0.8.2"
