--- auraescript_deno/Makefile
@@ -1,59 +1,0 @@
-# ---------------------------------------------------------------------------- #
-#             Apache 2.0 License Copyright © 2022 The Aurae Authors            #
-#                                                                              #
-#                +--------------------------------------------+                #
-#                |   █████╗ ██╗   ██╗██████╗  █████╗ ███████╗ |                #
-#                |  ██╔══██╗██║   ██║██╔══██╗██╔══██╗██╔════╝ |                #
-#                |  ███████║██║   ██║██████╔╝███████║█████╗   |                #
-#                |  ██╔══██║██║   ██║██╔══██╗██╔══██║██╔══╝   |                #
-#                |  ██║  ██║╚██████╔╝██║  ██║██║  ██║███████╗ |                #
-#                |  ╚═╝  ╚═╝ ╚═════╝ ╚═╝  ╚═╝╚═╝  ╚═╝╚══════╝ |                #
-#                +--------------------------------------------+                #
-#                                                                              #
-#                         Distributed Systems Runtime                          #
-#                                                                              #
-# ---------------------------------------------------------------------------- #
-#                                                                              #
-#   Licensed under the Apache License, Version 2.0 (the "License");            #
-#   you may not use this file except in compliance with the License.           #
-#   You may obtain a copy of the License at                                    #
-#                                                                              #
-#       http://www.apache.org/licenses/LICENSE-2.0                             #
-#                                                                              #
-#   Unless required by applicable law or agreed to in writing, software        #
-#   distributed under the License is distributed on an "AS IS" BASIS,          #
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.   #
-#   See the License for the specific language governing permissions and        #
-#   limitations under the License.                                             #
-#                                                                              #
-# ---------------------------------------------------------------------------- #
-
-default: install
-all: install
-
-cargo         =  cargo
-executable   ?=  auraescript
-
-compile: ## Compile for the local architecture ⚙
-	@$(cargo) clippy
-	@$(cargo) build
-
-install: ## Build and install (debug) 🎉
-	@$(cargo) clippy
-	@$(cargo) install --debug --path .
-	#sudo -E cp -rv examples/bin/* /usr/bin
-
-release: ## Build and install (release) 🎉
-	@$(cargo) install --path .
-
-test: ## Run the tests
-	@$(cargo) test                # Tidy output
-	#@$(cargo) test -- --nocapture # Full output
-
-clean: ## Clean your artifacts 🧼
-	@echo "Cleaning..."
-	@cargo clean --package $(executable)
-
-.PHONY: help
-help:  ## Show help messages for make targets
-	@grep -E '^[a-zA-Z_-]+:.*?## .*$$' $(firstword $(MAKEFILE_LIST)) | sort | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[32m%-30s\033[0m %s\n", $$1, $$2}'
