--- auraescript_deno/src/lib.rs
@@ -1,99 +1,0 @@
-// TODO: setup lints
-#![allow(dead_code)]
-
-use anyhow::{anyhow, bail, Error};
-use deno_ast::{MediaType, ParseParams, SourceTextInfo};
-use deno_core::futures::FutureExt;
-use deno_core::{
-    resolve_import, Extension, JsRuntime, ModuleLoader, ModuleSource,
-    ModuleSourceFuture, ModuleSpecifier, ModuleType, OpDecl, RuntimeOptions,
-};
-use std::pin::Pin;
-use std::rc::Rc;
-
-mod builtin;
-mod runtime;
-
-pub fn init() -> JsRuntime {
-    let extension = Extension::builder().ops(stdlib()).build();
-
-    JsRuntime::new(RuntimeOptions {
-        extensions: vec![extension],
-        module_loader: Some(Rc::new(TypescriptModuleLoader)),
-        ..Default::default()
-    })
-}
-
-fn stdlib() -> Vec<OpDecl> {
-    let mut ops = vec![];
-    ops.extend(runtime::op_decls());
-    ops
-}
-
-// From: https://github.com/denoland/deno/blob/main/core/examples/ts_module_loader.rs
-struct TypescriptModuleLoader;
-
-impl ModuleLoader for TypescriptModuleLoader {
-    fn resolve(
-        &self,
-        specifier: &str,
-        referrer: &str,
-        _is_main: bool,
-    ) -> Result<ModuleSpecifier, Error> {
-        Ok(resolve_import(specifier, referrer)?)
-    }
-
-    fn load(
-        &self,
-        module_specifier: &ModuleSpecifier,
-        _maybe_referrer: Option<ModuleSpecifier>,
-        _is_dyn_import: bool,
-    ) -> Pin<Box<ModuleSourceFuture>> {
-        let module_specifier = module_specifier.clone();
-        async move {
-            let path = module_specifier
-                .to_file_path()
-                .map_err(|_| anyhow!("Only file: URLs are supported."))?;
-
-            let media_type = MediaType::from(&path);
-            let (module_type, should_transpile) = match MediaType::from(&path) {
-                MediaType::JavaScript | MediaType::Mjs | MediaType::Cjs => {
-                    (ModuleType::JavaScript, false)
-                }
-                MediaType::Jsx => (ModuleType::JavaScript, true),
-                MediaType::TypeScript
-                | MediaType::Mts
-                | MediaType::Cts
-                | MediaType::Dts
-                | MediaType::Dmts
-                | MediaType::Dcts
-                | MediaType::Tsx => (ModuleType::JavaScript, true),
-                MediaType::Json => (ModuleType::Json, false),
-                _ => bail!("Unknown extension {:?}", path.extension()),
-            };
-
-            let code = std::fs::read_to_string(&path)?;
-            let code = if should_transpile {
-                let parsed = deno_ast::parse_module(ParseParams {
-                    specifier: module_specifier.to_string(),
-                    text_info: SourceTextInfo::from_string(code),
-                    media_type,
-                    capture_tokens: false,
-                    scope_analysis: false,
-                    maybe_syntax: None,
-                })?;
-                parsed.transpile(&Default::default())?.text
-            } else {
-                code
-            };
-            let module = ModuleSource {
-                code: code.into_bytes().into_boxed_slice(),
-                module_type,
-                module_url_specified: module_specifier.to_string(),
-                module_url_found: module_specifier.to_string(),
-            };
-            Ok(module)
-        }
-        .boxed_local()
-    }
-}
