--- auraescript_deno/src/runtime/mod.rs
@@ -1,10 +1,0 @@
-tonic::include_proto!("runtime");
-
-macros::ops_generator!(
-    runtime,
-    CellService,
-    allocate(AllocateCellRequest) -> AllocateCellResponse,
-    free(FreeCellRequest) -> FreeCellResponse,
-    start(StartCellRequest) -> StartCellResponse,
-    stop(StopCellRequest) -> StopCellResponse,
-);
