--- auraescript_deno_macros/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "auraescript_deno_macros"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-heck = "0.4.0"
-proc-macro2 = "1.0.44"
-quote = "1.0.21"
-syn = { version = "1.0.101", features = ['full'] }
