+++ auraescript_macros/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "macros"
+name = "auraescript_macros"
 version = "0.1.0"
 edition = "2021"
